Fix the input lag issue
This was relatively minor but I wanted to make a definite note of it
so that I can remember it for posterity.

The "input lag" I was experiencing was not a frame/speed issue, like I
thought. Instead, Camera2D (by default) has two "Drag Margin"
properties. This makes the camera behave more like Mario, or something
equivalent.

Of course, this isn't what we want, so I've switched it off.