Reintroduce the Edge Determine algorithm
As part of the effort to recreate Prototype I, I have readded the
accursed Edge Determination algorithm into the WorldBase class. My
plan is for each "type" of world to inherit from the base world; the
key difference between each world being the generation processes.

One of the key changes to the edge determine algorithm is the addition
of extra detail layers for smoother tile transitions. Unlike in
Python, this hasn't caused an en-masse drop in frames. This has made
the algorithm messier in spots (cough cough bad sort), but has also
gotten rid of some of the weirder quirks.

Because of the increased speed of Godot, it's now more feasible for us
to update the edge tiles in real time. This will enable us to add
things like snow buildup, or grass giving way to dirt. For that
reason, we now have an edge determine function that works on a single
tile, rather than the edge pass.

Finally, we added a priority dict in lieu of being able to assign
properties to tiles. That feature is supposedly coming, but I'll
admit to being irritated that it's not here already. This could
severly hamstring our development...