Begin adding new world types
We'll want to add new world types. However, I want to utilize
inheritance while doing so. Since there aren't any mixins, I'm going
to create a world type for each feature. In addition to supplying a
generic function to create that feature, it will also provide a
generate function that serves as a sort of "Feature Zoo" demonstrating
test cases for that particular generation technique.

We've already shown this off with the edge determine, next we're going
to take on rivers. That, however, will require it's own
extra functionality...