Add Bresenham functionality
To paint properly ebbing and flowing rivers, we need to be able to
paint in circles, and along lines.

To do this, I've ported over our Bresenham code from the Python
project.

Unfortunately, Godot doesn't allow functions to be passed as
first-class objects. It also doesn't have varargs, apparently. So we
had to change how arguments are passed to the BresenHam functions.

Godot does offer untyped arrays, and a sort of pseudo function
passing. You can create a FuncRef object that allows you to invoke
a specific function of a specific object at any time. Handy!

Although I typically abhor the what I call the "Do-er" pattern, where
an object must specifically be instanced just to perform an action
that could have been served perfectly by a regular library full of
functions.

However, I've recognized that nearly everything about Godot pushes you
towards the "Do"-er pattern; everything is a node, and each node is a
class. So, while I feel like the structure is over-engineered and
unnecessary, I've yielded in this instance.

Two of our "library" scripts have now been attached to nodes:
TileData.gd (added to BaseWorld) and Bresenham.gd (added to
RiverWorld). Again, I'm not a fan, but it does make it quite clear
which scenes are invoking what libraries.