Figure out an inheritance order for our worlds
Okay, so originally I thought that I would do the "OpenSimplex" world
last in my list of worlds, because surely that's the most complex one.

Then I realized, my river algorithm requires noise input; so I'd have
to make our "OpenSimplexWorld" FIRST.

Then, while trying to work on that, I realized that the
OpenSimplexNoise in Godot was a resource, not a Node like I was
expecting. Each instance of something that requires OpenSimplex
generation will have to instantiate it's own generator.

I also realized that, although generating the ground cover is the
"first" step in the generation process, it's really the one that's
going to be the most specific to whatever our world type is. Rivers,
on the other hand, will always want to follow the same sort of
algorithm and appearance. It makes sense to have an inherited world,
and separate script, just for rivers.

So I reverted back to the way things were at the start. However, the
River generator is now a RiverPaint node, and has a configurable noise
generator. Any subsequent worlds with river generating needs should be
able to configure this for their inherited scene appropriately.

Now, we actually need to port over the lattice-based river generation
from Prototype I.