Add in very basic Simplex nonsense
After a long time stewing on how to paint rivers, I decided that I had
better things to waste my time on. Like actual mechanics, for example.
But I am who I am, and couldn't bring myself to test mechanics against a
blank field.

I had a bit of a stroke (of genius?) and conceived a system where the
ground was painted using the intersection of two simplex noises: the
moisture and the soil. The idea was these two values in combination
could also satisfy generating detail tiles and forests as well.

The result is... ok. Not great, not terrible. But it's enough for me to
move on from, for now.