Implement boid-death
In order to implement the collision-avoidance feature of our boids, I
decided to create a punishment for collision. It plays out a small
explosion animation on contact with an obstacle and removes the Boid -
for right now, the removal is simple visual. We'll focus on removing
the node in subsequent, death focused updates.