Add basic avoidance
Added in basic avoidance/pathfinding. Basically, it always scans an
area in front of it and chooses the first unblocked path. Sort of a
greedy, path of least-resistance kind of thing.

There are two issues that I currently see. Firstly, the boid kind of
automatically jerks whenever they turn, since we set the rotation
instantly. That's nothing too difficult, just need to introduce
gradual rotation to the boid's _process() function.

More concerning is the second issue - the boid, flying dead ahead, is
clipping on obstacles and dying. It seems that the greedy algorithm's
single-minded nature is destroying the boid. We need to evaluate our
surroundings and somehow weigh the inputs into a cumulative direction.