Make better directional decision making
In the last commit we implemented an obstacle avoidance algorithm
where our Boid would raycast in front of itself and choose the first
non-colliding path that was available.

When the boid had nothing in front of itself, it only had to do one
raycast. However, the Boid does not exist as a singular point - it has
a width. So the boid was wizzing past obstacles, clipping them on the
back half, and then exploding.

I tried several different ideas using vague memories of math and
averaging out values, but these generally only contributed to the Boid
gleefully pitching itself into a wall and exploding. The main source
of my consternation was that we were combining left and right inputs,
which we then averaged out. This came to nothing and created a Boid
who wouldn't turn.

After googling around, I realized I wasn't going to abracamath the
Boid into not killing itself. I had to cut the problem into two
halves - left and right. We would need to process these two halves
separately and then compare the outcomes. I also realized that we
needed a DEFINED center where the Boid is in danger of of dying if
ANYTHING is in that area.

This gives us four tiered zones to watch, which we can then act upon
in a strict order.

Just to make sure all of this works, I also added a bunch of debug
options, including collision sprite markers capable of cleaning
themselves up.

My next concern is about our deliniation of the "center" which is
currently angular but I think we will need to change so that it can be
easily adjuststed based on the Boid's width.