Turn the boid gradually, not instantly
The boid used to instantly rotate to the calculated "ideal". I felt
this was good for initial establishing the ray-casting algorithm, but
the time has come to replace it.

The boid now turns gradually to avoid danger. The boid no longer
identifies an "ideal" vector to rotate to - instead, it identifies an
ideal direction to rotate to. It will then rotate in that direction
until the way is clear. The idea is that, with _physics_process(), the
Boid will naturally rotate itself out of danger.

The new issue is that the boid keeps killing itself anyway, since it
can no longer turn out of the way in time. What we need is to check
farther in front of us depending on our speed; we'll have to change
our left-right-center processing in kind.