Update L-R-C decision pathing
Updated the left-right-center pathfinding decision making in direct
response to our now gradual turning.

When performing our collision raycasting, we now cast the points
farther in front of of us depending on the boid's current speed - the
faster the boid, the farther the cast. The turning speed is likewise
reduced as we speed up.

The current issue is that the boid always go full speed and so
constantly kills itself - we'll have to make a boid that intelligently
slows down and speeds up.

Also, since it was bugging me, our boid now ACTUALLY dies when it
explodes - it stops moving instead of becoming a ghost. It also
doesn't disappear, it just turns pitch black. As a test, I tried
translating the artifical linear & angular velocity into their physics
engine equivalents. Now when the boid dies, it does an entertaining
spin-out.