Break path-blocked ties using counts
A consistent issue with the boids has been tie-breaking between
equally palatable choices. I was noticing that, in one particular
phenomenon, one of the sides and the center would be completely
blocked, while one side was partially blocked.

Since the pathing algorithm previously had an all-or-nothing attitude
towards blocking, the AI had no where to go and flew straight into
death.

With this update, we evaluate directions by counting the number of
collisions in each direction. In the event of a tie, we then use this
data to choose the least blocked path. This results in much less
death over previous algorithms.