Put the Boids through more rigorous testing
I decided to try pushing the Boids through more rigorous testing - so
I added more boids and a new class of bouncing balls to test the boids
against randomly moving objects.

And, ooh boy, they did poorly. My current favorite behavior is when
the boids attempt to avoid each other and do a mathematically perfect
ballet kamikaze.

It seems that my pathing algorithm was really only good at avoiding
static objects - I'll have to adjust it to compensate for the
colliding objects and their vectors. This is gonna be hard...