Change to a force-and-torque driver
So I realized that I needed to modulate how fast the boids were going
in order to make them better avoid incoming danger. And I decided, for
some reason, that the best way to do that would be with force and
torque.

It made sense at the time. I think.

The new pathing algorithm now assigns an "ideal" direction for the
boid by adding together all the projected obstacles. It then attempts
to align to this direction using force and torque.

The avoidance algorithm works pretty well; the torque-and-force does
not. So I will correct that in further updates.