Adjust to multitudinous quantities of BOIDS
After settling on an avoidance algorithm, now comes the other key
aspects of the Boids algorithm: cohesion, and alignment. To build out
those parts, I added a crap ton of Boids to the simulation.

I first had to increase the raycast count to 35 - the Boids were so
small that they were slipping through each other's raycast nets.

In my vision of this prototype's final iteration, friendly Boids will
not collide with each other. So I changed around some of the collision
layers and masking - Boids cannot, currently, physically collide with
one another. Since we have to change the masks around at death (which
I think will rapidly get out of hand), I isolated the dieing tasks in
their own function, die().

However, I noticed that the Boids were still dieing when touching each
other - that's because the body_entered signal triggers WHENEVER a
body is entered. It doesn't matter whether the physics COLLIDE with
that body - a body entered is a body entered. To counteract this, I
came up with the idea of a boid_team, which we'll use to code
friend/foe for the Boids. I'm not sure how we'll execute it exactly -
right now it's just a string and that's good enough.

I also separated all raycast/threat-vector code from _integrate_forces
into it's own function, calculate_avoidance_vector. This better
represents the functions place in the avoidance-alignment-cohesion
paradigm.

But now I'm thinking - what if we separated out the brain from the
boid? Then we could make just about anything a boid...