Separate boid behavior from boid type
I wanted to - eventually - have more than one sort of boid. Massive
boids, microboids, longboids, shortboids. My problem was that the
core boid code - the driving and avoidance - was tied to a specific
boid, the Microboid.

To implement different types of boid, I'd need to set-up the boid code
so it could readily be added into a scene. A plug-and-play behavior
generator.

Elements relating to drive-and-detect were split off from Boid_Micro
to form their own scene, Boid_Brain. And it's pretty easy to plug in -
just takes a couple of function calls.

The only element that wasn't so plug-and-play is the Boid's team. It
was easy to implement when all boids were the same scene, but now the
detection exists at a level below the boid. We'll need to come up with
some sort of configurable scheme...