Create a faction node to manage relationships
Previously, a boid's team (and whether a collision would be lethal or
not) was determined by comparing a 'team' string that each boid had.

I wanted to replace the 'team' variable with something better - this
new 'team' concept would manage the relations between itself and other
such team objects, and would need to be accessible from the all levels
of a Boid. As I figure, we currently have two levels:

	Boid -> BoidBrain

In the future, this will expand to at LEAST three levels:

	BoidManager -> Boid -> BoidBrain

The new 'team' concept would need to permeate all three levels and
provide information about hostility, neutrality, and friendliness.

This has been implemented as BoidFaction. As the simulation currently
stands, we only really have the two levels of Boid and no manager
class as of yet. Ergo, the BoidFaction sits in BoidSpace and is
manually passed down to each Boid. In the future, whatever class is
responsible for creating the Boids will perform this passdown
procedure.

Since the 'team' concept no longer relates to the collision layers and
masks, the change_team function in BoidBrain.gd has been removed in
favor of two functions that assert the Boid's state: set_mode_boid &
set_mode_obstacle.

I also recognized a need for more "cleanup" of dead boids (i.e. saving
processing by turning off monitoring), so I implemented that as well.