Add alignment and cohesion vectors
With the avoidance vector in a more-or-less finalized state, and the
boid team resolution (temporarily) implemented, the next natural step
was to implement the remaining two components of boid algorithm:
alignment and cohesion.

Forunately, this was much easier to set up than the avoidance vector.
Since these two vectors refer to flock mates, and the BoidBrain
passively tracks flock mates, it was just a matter of sorting through
the flock to get the information we needed.

The result was a bit too powerful for it's own good, I'm afraid. The
boids would enter this weirdo spiral dance that they couldn't escape
from. That was due to the cohesion vector being too powerful, and
sampling all the boids - forward, and back.

To help alleviate this issue, I've manually scaled each vector by
different amounts. Avoidance & alignment have very nearly the same
value; only cohesion has been more or less neutered.

To me, this indicates a problem with the algorithm. Either how the
boid controls or how we sample for cohesion will need to change. Given
how PAINFUL getting the boid movement correct was, I'm more inclined
to change the sampling.