Only practice cohesion with forward boids
As stated in the last commit, I had to put weights on the various
Boid algorithms to make sure everything was working correctly.

To create a "better" algorithm, I've added a separate detection area
just for calculating cohesion vectors. This new shape only covers an
area "forward" of the boid, which prevents the doubling-back
phenomenon we saw earlier.

Since the new cohesion detection shape needed to be manually input,
I've removed the RAYCAST_MAGNITUDE as a GUI configurable item. It is
now a constant, so that we don't accidentally change a fundamental
component of the simulation.