Spawn boids & balls on command
To finish up the boid experiment, I want to readily play around with
the amount of boids and the amount of obstacles (death balls).

To that end, I've created the BoidSpawner scene, which now
encapsulates the BoidFaction scene. This scene listens for input on
the left & right mouse clicks; it then spawns a boid or a ball.

To make the control scheme clearer, I also added a small GUI. It's a
of text lines to tell the user the control scheme.