Adjust scalars on the fly for easy experiments
I was trying to fine-tune the different scalars for our 3 input
factors, when I realized that it would be an ideal demonstration tool
to actively change these values on the fly.

To facilitate this, I had to add a global/static node - in Godot, this
is done with the engine's AutoLoad functionality. This new node,
BoidGlobals, contains the code for adjusting the scalars in response
to inputs as well as the scalars for themselves.

I also added some GUI text so that I don't have to remember the
controls off the top of my head. I wish I knew a way to make them
dynamic - i.e. read out the configure control and list it. Possibly
even with some nice looking buttons. Oh well; that's a wishlist for a
future project, another time.

With this commit, I'm considering the Boid proto-project a success!
Hooray!