Remove the old "world" prototype
As we move on to a new prototype, there's no need for us to continue
holding on to the "world" prototype. To be quite honest, I'm thinking
that we may end up doing something radically different for our terrain
generation. It was a fun experiment, but I have better things to work
on.

The boids, however, will get to stay - this next prototype will depend
on the lessons learned from that particular prototype.