Add a point-and-click "muster flag"
I want to make this next prototype about moving units - groupings of
self-driving entities around. To act as a foci for the group, I'll be
(trying) to use command flags. They'll function as they do in Total
War or Black & White 2 (my main inspirations, really).

After some experimentation, I created a basic muster flag (just a
sprite for now) and managed to move it around by clicking. I'm not to
satisfied with how our processing is spread across two separate
functions - since we cannot poll the physics space from our input
function.

I also found something quirky - calling the look_at function on a node
seems to reset the scale of the node? Not ideal, all things
considered. I was trying to force the flag to look at the camera, but
it was messing with the scale. Fortunately, I realized that an
orthographic camera doesn't need the node to face the camera exactly -
just the general direction.