Play with Godot 3D & create a Unit scene
I've been messing around with Godot's 3D components - especially
materials, as those are a relatively new concept to me. You may notice
the unecessary amount of new textures and material resources. All part
of the process!

I also messed around with lighting - I added a nice, pink cube that
slowly but surely spins around. It does nothing and it mesmorizes me.
The FPS is stable, and the cube casts a shadow below it. Might seem
like an average (not important) detail, but shadows are very important
for gauging the position of objects - and THAT is very important in an
isometric environment.

Now, for something actually relevant to our development path: this
prototype will have groups of NPCs called "Units" - the individual
NPCs will be called Pawns, Mobs, or Bugs (possibly all 3). The Muster
Flag from the last commit was intended to show the anchor/rallying
point for the unit.

After playing around with it, I found that a sprite floating in space
made a grossly inadequate expression of position. So I added a sort of
placement-specifying cube... thing. That now moves when right
clicking. It does a better job of communicating where the unit is
going to go then the sprite did.

I also created a UnitPawn scene that will be the Mobs/Pawns/Bugs for
the Unit. Right now it's just a sprite and and a collision shape.

Next, we'll focus on making those Mob-Pawn-Bugs actually animate.