Move the UnitPawn some orders
The UnitPawn was inert - but now, no longer! I have created a very
basic driver for the UnitPawn scene.

It works on the same fundamental concept as the Boids - we have a
target destination, and we move towards that destination. Currently,
this target destination takes the form of a move_order vector.

This presents several questions - how is the move_order set? How
frequently does it change? What determines if it needs to be
changed? When do we decide if it's done? How does information flow
from the unit to the pawn and back again? What information would
that be, anyhow?

All this, and more to decide! For now, though, I think I will focus
on making the Pawn-cylinder not fall over and freak out. Much as I
love a good freak-out...