Fix the UnitPawn physics freak out
Previously, our UnitPawn object was animated using a RigidBody object.
I was using this based on a hazy memory that this was appropriate for
NPCs. As it turns out, that's for 2D. Or it's a lie. Either way,
that's just not true in 3D.

Trying to animate (i.e. move) a 3D RigidBody is a bit stupid. The body
falls over, spins, and generally freaks out as you keep forcing the
position and making the physics engine cry. It cries small tears, but
cries all the same.

The solution was to change UnitPawn to a 3D KinematicBody.
KinematicBodies have more low-key physics simulations - just some
basic collisions. It doesn't even have built-in gravity.

Even after implementing the KinematicBody, I noticed that the UnitPawn
freaked out a bunch. This was because it was trying to hew too closely
to the follow-path and kept over-shooting left and right. Rapidly.
Adding some give (in the form of PATHING_TOLERANCE) fixed this issue.

While I was making PATHING_TOLERANCE a constant, it only made sense to
constant-ize the tolerance we had for our goal point. And behold - the
GOAL_TOLERANCE constant has been made!