Experiment with UnitPawn shadows
I had some ideas for different lighting effects to use in this
project, but I had no idea how the current UnitPawn design would look
with various forms of lighting. We had some DirectionalLighting in our
main scene, but that was it.

Also, the lack of depth (a direct side effect of the isometric camera)
made gouging the actual position of the UnitPawn difficult. If we
could get the UnitPawn to cast it's own shadow, that would help us
with showing the depth.

Initially I tried using a shadow-only cylinder that inhabited the same
space as the hitbox. This worked well for casting an (outsized)
shadow, but blocked light to the UnitPawn's core sprite, casting it in
darkness.

After some experimentation, I found that a 3D sprite could cast a
shadow when the Override Material was set to a Billboard Mode and
the Depth-Draw mode was set to Opaque Pre-Pass. I set it to
Y-Billboard mode, which locks the sprite so that it always faces the
camera and rotates only on the Y-axis. This was exactly what we wanted
the sprite to behave, so that's a win.

However! It either couldn't, or didn't, cast a shadow from a
DirectionalLight. I had to add a WorldEnvironment to test the scene in
darkness - with spotlights and omnilights, we could cast some really
good shadows.

Messing around with the WorldEnvironment, I found that setting the
Ambient Light Energy could control the "severity" of the shadows.
Using this, I was also able to determine that the 3D Sprite can also
accept shadows, especially those from DirectionalLight.

However, the UnitPawn still wasn't casting a shadow. I tried messing
around with a decal plugin, but it seemed like there was no way to
control the size of the decal.

Finally, I settled on a flat shadow sprite that's connected to the
UnitPawn's bottom. We'll have to find some way to glue the shadow to
the ground...