Cast true shadows with the UnitPawn
The entirety of the last commit was just trying to get the UnitPawn's
core sprite to just cast a shadow. Any shadow! The commit itself was a
concession of defeat, and a documenting of what I had learned.

But I was REALLY not happy with my solution. That dopey little shadow
I added was just gross - SURELY there was a better way. So, I was
experimenting with hiding objects behind the sprite to cast a shadow.

That's when I spotted the sprite ACTUALLY CASTING A SHADOW. It was
only casting a shadow while at its apex; the shadow seemed to be
slipping into a void space once it got close enough to our floor.

After some googling, it seems like my problem was caused by the normal
values of the DirectionalLight's Directional Shadow and Shadow
properties. The normal adjusts shading on surfaces; it was too high so
it was cutting the shadows out. Or something. I'm still not clear on
how it works, but the point of the 3D project is to learn.

After some playing around, I found that if I moved the Sprite slightly
off the floor, turned off billboard mode (it was causing render
errors, not that the billboard mode was making that much of a
difference anyway), and shifted the sprite upwards slightly so it was
floating just a little bit.

One thing that I did notice - when the camera was in perspective mode,
the low normal values on the DirectionalLight made it so that the
ground-plane was self shading. This banded the plane with weird little
gradient shadows. This doesn't appear in orthogonal perspective, so no
big deal.