Order the UnitPawn through the Unit node
In previous prototypes, we ordered the UnitPawn to move somewhere by
reaching from our main scene (ParadeGround) down to the Unit scene,
then directly (and manually) manipulating the appropriate variables.
Obviously that wasn't sustainable - we'd need a better solution.

And that solution is here - or it's still on it's way and this is just
an interim step! Either way, things are different now.

The idea is this - a Unit receives a move order via the brand new
order_move() function. The Unit emits a move_ordered signal,
broadcasting the new target-position for the Unit. The UnitPawn, which
has hooked into the move_ordered signal through the register_to_unit()
function. The UnitPawn then calculates where to move by calling back
to the Unit via the get_pawn_index_pos() function. This function
calculates where a given UnitPawn should go, which we then pass to
the UnitPawn's set_target_position() function.

My hope is that looping things through get_pawn_index_pos() will allow
us to easily set up formations at the UnitPawn level. Or maybe it
won't; we'll see.

The set_target_position() function is currently pretty dinky. However,
I kept it separate so that we have a way to order around UnitPawns
that are individual/not in a unit. It might grow, it might not. We
shall see!

In the meantime, I added a bunch of UnitPawns to the base Unit. This
allowed me to test the UnitPawns moving together as an actual Unit.
It's iffy - patchy, even. The biggest issue is that they are colliding
with each other, and can't really re-orient themselves within the
Unit. We might have to see if there's a way to fix/mitigate that.

Also, I noticed that when ordering the Unit around, I was actually
selecting the UnitPawns themeselves, causing the target to float.
That's not what we want! We'll have to fix that.