No-collide the UnitPawns of a Unit (maybe?)
I want each Unit to be a large-scale collection of UnitPawns; with
anywhere from 4 to 32 UnitPawns, depending on the size. I may also
want them to rapidly self-organize, or re-organize to face new
threats (on the flanks, or the like).

However, this would be difficult to achieve if all the UnitPawns are
colliding with each other. I briefly contemplated using collision
layers to manage this (which seemed to be the biggest recomendation).
However, after I added some basic layers (terrain, obstacles, etc), I
only had 17 layers left. This capped me at a maximum of 17 units. That
could be too many for what I want to do, but I didn't want that
restriction.

Instead, I added an extra function to UnitPawn:
no_collide_with_list(). This function adds collision exceptions for
each node in the list. This function then needs to be called on every
node with a list containing (at least) every OTHER node.

While this is an N^2 complexity solution (BAD!) we do have a couple of
saving graces - first, we should be capped at 32^2 (1024) iterations.
Secondly, this (currently) occurs when the Unit node first enters the
scene. We can change this to be part of the spawning process that
would occur behind a loading scene in a normal game.

However, I felt like "units do not collide with each other" could be a
pretty major design decision (one that I'm about 85% committed to), so
I added a boolean configurable to toggle collision on.

While I was updating the Collision layers, I took a second to update
the intersect_ray() raycast that gives the Unit its "orders". We now
pass the function call an empty array (node exclusion list), and a
collision mask. The empty array was just an optional argument we HAD
to include to get the collision mask. The collision mask is an int
value; fortunately, we only need to collide with bit 1, the Terrain
layer.