Change the UnitPawn's animation per direction
This commit implements some basic animation for the UnitPawn. The
current animation is swapped in real time depending on the UnitPawn's
horizontal velocity.

So the UnitPawn's sprite will always face the camera; we just have to
swap around the animation and the sprite resource. I'm 50/50 on this -
I like the forward facing sprites, but I also want them to really
EXPRESS their flatness. It almost feels like I'm hiding the flatness
of the sprites, where as what I really wanted to do was show them off
(I guess?). I'm sure the art style will continue to develop - and who
knows, maybe an actual artist will give me some feedback sometime.

The actual algorithm to calculate direction-facing was pretty simple.
I was unsure about where the calculation should actually occur; I
ultimately decided to update it in the frame loop (_process) as it was
a visual update. Visuals in the visual loop, physics in the physics
loop.

One thing I noticed: I need to adjust the GOAL_TOLERANCE because all
of the UnitPawns were stuck in a perpetual move state. It seems they
can never get closer than 0.056 units away from their goal. So I set
it to 0.057. That's a good band-aid, but we should see what the issue
is and if we can fix it.

I currently suspect it has something to do with how and where we
measure the distance from - we need to measure from the UnitPawn's
feet rather than the center. But that's just a guess.