Spawn in UnitPawns at runtime
Previously, a Unit's UnitPawns were preset and packed as part of the
scene. This was good for testing but would ultimately be bad for
deployment. I plan to re-use the Unit scene for ALL units, so it needs
to be multi-purpose. That means spawning in the appropriate kinds of
UnitPawns on the fly, rather than going with a fixed set.

To fix this, I created new UnitPawn scenes that inherited from the old
scene. These new scenes adjust the VisualSprite to one of the 8
available. We then PRELOAD these scenes into the Unit script and spawn
the UnitPawns using a round-robin-random.

It's important to note that we load the scenes using preload(), rather
than just load(). This means the nodes are included at compile time
and can run faster. Still, there is a definite performance hit we took
at start-up. There wasn't any noticeable dip having the UnitPawns just
move around, though.

Thinking about it now, we may want to actually create packed scenes of
different Unit types, just like we did for UnitPawns. We can specify
the UnitPawns, and even mix unit types. That might just be the thing
to do...

Just some final notes: we now determine the "per-row" count of
UnitPawns using an ugly formula that's actually pretty simple - just
rounding up the square root, which should give us the nearest value to
fit everything in a square. At least, it will fit everything in the
Unit's zone. To go with that, the size of the Unit is now determined
by the UNIT_INITIAL_STRENGTH value.