Perform UnitPawn goal-checking from the base
Previously, we checked the UnitPawn's position against the goal by
using the UnitPawn's global origin/position/translation. This worked
pretty well, however I noticed that the UnitPawns never got much
closer than 0.056 units.

The issue was that the origin/position/translation was the MIDDLE of
the UnitPawn. It was, thusly, a FIXED DISTANCE from the floor. It
would never be closer to the goal (which was on the floor) because it
was blocking/colliding itself.

The fix, then, was to measure the distance to the goal from the bottom
of the UnitPawn - where the collision model meets the floor. The
UnitPawn is now having a much better time reaching it's goals. It's
not perfect, so we still need the TOLERANCE constants.

On a side note, I shrank UnitPawn down to it's original size of .1
units. Having the sprites floating on nothing was bugging me.