Move the TargetGroup of Unit via global transform
Previously, we were moving the TargetGroup of unit (the positional
or ordered indicator) via the node's translation value. However, this
caused an issue - if we moved the Unit from (0,0,0), the TargetGroup
would always be offset by that vector.

However, I noticed that the UnitPawns themselves appeared in the
appropriate positions. Looking through the code, I saw the Pawns
checked global_transform.origin to get their current position.

Manipulating this value, instead of translation, fixed the
TargetGroup. It put everything exactly where it needed to be.