Switch from checkerboards to color-grids
When I was younger, I made a lot of maps in Valve's Hammer Editor. I
was always a fan of their orange-white-grey color scheme they used for
map planning. It was so aesthetically pleasing that it frequently made
it into TF2 maps as it's own bonafide style.

And, while I liked the checkerboards, I wanted to pay tribute (even
debuggery-tribute) to those classical orange-and-grey grids.

So, I swapped to using some Valve style grids I cooked up. One
immediate advantage: the grids line up with the editor's built-in
grid. This makes it easier to use them for measuring (which I think
was the purpose of Valve's grids).

I started making a bunch of Valve-Grids in two different color schemes
and was making more, but realized that was excessive - there wasn't a
need for more than 3 grid materials.