Adjust how we calculate the 'current' angle
The 8-directional sprites were an improvement, but the sprites were
having trouble facing the right direction.

To diagnose the direction, I added a brand new debug item to the
UnitPawn - a directional arrow that we spin to match the set angle.

What I found was that our concept of the "current angle" was off.
The display sprites are intially rotated to 45 degrees - this meant
that our sprites were 45 degrees separated from the UnitPawn's true
angle. Subtracting those 45 degrees before the direction decision
allowed us to align appropriately.

The direction decision cascade was also not working - each section
needed to be shifted by 22.5 degrees (45/2), so that the sections
were appropriately aligned with the multiples of 45 (0, 45, 90, 135,
etc.)