Add a bear model
The original goal for this prototpye was to have a unit of meeples
fighting some sort of giant creature. I've always been fond of David &
Goliath type stories (except there's 36 Davids in this one).

I finally found a visual configuration for the soldiers that I was
happy with (even if it is vomit clowns), so that left the Beast. I
wanted it to be an obscenely large beast, and I wanted it to be a
Bear. Bears are big, fearsome, and powerful. And against clowns (why
clowns why) I thought it would be funny.

Only problem - isometric or 3/4 perspective bear sprites with 8 degree
motion is, APPARENTLY, too specific a category for open source game
art. Fair enough. But, what would I do then? Make my own? WHAT ARE
YOU, CRAZY?! I MADE THE CLOWNS! THAT SHOULD BE ENOUGH TO BAN ME FROM
ART PLANET!

So, if sprites wouldn't work, that left models. I have made models for
3D printing before, based on my lineart. Surely it couldn't be that
hard? Well, it wasn't that hard, but it did involve a whole art
pipeline.

First, I created a stylized bear off the California state flag using
GIMP. I exported this as a .bmp, which I then brought into Inkscape.
From Inkscape, we re-traced the image to create a vectorized version
and then exported to a .dxf file. This was then imported into OpenSCAD
where we extruted, translated, scaled, and rotated it. Now that SHOULD
have been the end of it but OpenSCAD doesn't export to ANYTHING Godot
compatible. So, I had to export as STL, import the STL into Blender,
re-scale and re-translate it, and then export as a .OBJ. THAT could
finally be brought into Godot.

The final result? I can see it. I don't care anymore. I can see it,
it's a bear, I'm good. I don't need anymore. I'm finding that the flat
isometric sprites - while a lot more effort in terms of assets - do
genuinely have a charm all their own. Too bad I got banished from ART
PLANET for CLOWN ATROCITIES.