Use SURT style terrain
So, I made those valve style grids. Bunch of colors, super nice
looking, all good. But, now that I have my clowns, and I have my bear,
I need to start making them do stuff. The clowns can move, but they do
so in a linear way. I need to start working on pathing, then I can
work on attacking (and then I can get this bear killing some clowns).

So, to me, obstacles means terrain. And terrain means it's time to
move away from this flat grid. And THAT means picking a more coherent
art style than magentas and purples.

We've been falling back on Surt's work a lot, and I fully intend to
keep doing that. So, I replaced the "generic" Valve grids with grids
colored using Surt's NES color pallette. The result is looking good,
so far!

Also, I adjusted how the textures were scaled. The grid was meant to
be standardized but it kept shrinking or growing depending on the size
of the attached mesh. I was able to get around this by enabling the
"World Triplanar" parameter flag and the UV 1 "Triplanar" flag. The
result is materials that don't move with the object - it looks really
weird, but works perfectly for what we want. At least for the grids.