Update UnitPawn size, collision, and position
There were a few issues I had with the UnitPawns that I wanted to
correct.

Firstly, the UnitPawns were centered such that the local node's
origin was in the middle of the Pawn. I thought this was pretty much
standard but it meant we had to shift the UnitPawn upward by half of
it's height everytime we spawned it in at a point.

To get around that, the UnitPawn has been shifted upwards so that the
local origin now originates from the Pawn's feet.

I also noticed that the UnitPawn sprites were scaled to .04. I
generally prefer to not mix scaling options (like scale and pixel
size) so I set those to full scale and readjusted the pixel size.
Although the result is good, I noticed that pixel size is NOT very
granular. We may need to go back to mixed scale if we run up against a
sprite we can't size correctly.

Finally, I updated the size/shape of our collision model, but only for
the Clowns and Doom Marines. Unlike the Surt sprites, the X-COM style
Clowns are thin (not boxy), so the boxy-style cylinders weren't
working. I created new collision models for the Clowns that more
accurately reflect their sprites. Also updated the Doom marines. We
don't use them, but I like to stay on top of things like that.